+++ src/schemas/clinical-mutations/commitRegistration.ts
@@ -1,14 +1,18 @@
-// commitClinicalRegistration: async (
-//   obj: unknown,
-//   args: { shortName: string; registrationId: string },
-//   context: GlobalGqlContext,
-// ) => {
-//   const { Authorization } = context;
-//   const { shortName, registrationId } = args;
-//   const response = await clinicalService.commitRegistrationData(
-//     shortName,
-//     registrationId,
-//     Authorization,
-//   );
-//   return get(response, 'newSamples', []);
-// },
+import { commitRegistration } from '../../submission/submission-to-clinical/submission-to-clinical';
+import { get } from 'lodash';
+
+const commitClinicalRegistration = async (
+  obj: unknown,
+  args: { shortName: string; registrationId: string },
+) => {
+  const { shortName: programId, registrationId } = args;
+
+  const response = await commitRegistration({
+    registrationId,
+    programId,
+  });
+
+  return get(response, 'newSamples', []);
+};
+
+export default commitClinicalRegistration;
