+++ src/schemas/utils.ts
@@ -19,6 +19,7 @@
 
 import get from 'lodash/get';
 import {
+  ActiveRegistration,
   SubmissionValidationError,
   SubmissionValidationUpdate,
 } from '../submission/submission-entities';
@@ -72,6 +73,36 @@ const convertClinicalFileErrorToGql = (fileError: {
     message: fileError.message,
     fileNames: fileError.batchNames,
     code: fileError.code,
+  };
+};
+
+const convertRegistrationDataToGql = (
+  programShortName: string,
+  data: {
+    registration: DeepReadonly<ActiveRegistration> | undefined;
+    errors?: RegistrationErrorData[];
+    batchErrors?: { message: string; batchNames: string[]; code: string }[];
+  },
+) => {
+  const registration = get(data, 'registration', {} as Partial<typeof data.registration>);
+  const schemaAndValidationErrors = get(data, 'errors', [] as typeof data.errors);
+  const fileErrors = get(data, 'batchErrors', [] as typeof data.batchErrors);
+  return {
+    id: registration?._id || undefined,
+    programShortName,
+    creator: registration?.creator || undefined,
+    fileName: registration?.batchName || undefined,
+    createdAt: registration?.createdAt || undefined,
+    records: () =>
+      get(registration, 'records')?.map((record, i) => convertClinicalRecordToGql(i, record)),
+    errors: schemaAndValidationErrors?.map(convertRegistrationErrorToGql),
+    fileErrors: fileErrors?.map(convertClinicalFileErrorToGql),
+    newDonors: () => convertRegistrationStatsToGql(get(registration, 'stats.newDonorIds', [])),
+    newSpecimens: () =>
+      convertRegistrationStatsToGql(get(registration, 'stats.newSpecimenIds', [])),
+    newSamples: () => convertRegistrationStatsToGql(get(registration, 'stats.newSampleIds', [])),
+    alreadyRegistered: () =>
+      convertRegistrationStatsToGql(get(registration, 'stats.alreadyRegistered', [])),
   };
 };
 
@@ -211,6 +242,7 @@ const convertClinicalSubmissionUpdateToGql = (updateData: UpdateData) => {
 
 export {
   convertClinicalRecordToGql,
+  convertRegistrationDataToGql,
   convertRegistrationErrorToGql,
   convertClinicalFileErrorToGql,
   convertRegistrationStatsToGql,
